requirements-dev.txt
# Local development dependencies for My Story.
#
# These are ONLY for running tests and working on the code locally.
# They are NOT bundled into the Lambda zip at deploy time — each Lambda's
# own requirements.txt controls what ships to AWS.
#
# Install (inside an activated virtualenv) with:
#   pip install -r requirements-dev.txt

pytest>=8.0
pytest-cov>=5.0
moto>=5.0
boto3>=1.34
responses>=0.25
python-jose[cryptography]==3.3.0
cryptography>=41.0.0
